issue_16: コンポーネントに分けファイル分割
